Implementationof ConcertIOException
Now a concert will throw an error that the MainGUI class will catch if any incorrect information is detected from file when the application is started. A message will appear detailing to user what specific concert failed to load from file, telling them which line the concert is on in the Concert_list file so they can correct it.