Major JavaFX Update
Massive overhaul of GUI components and internal structure.
The application now uses an MVC framework, with controllers dictating the GUI visuals, responding to user events, and maintaining the concerts in memory. The views are fxml files, and they all point to the EventController class to respond to user actions in any of the views. 
The models are all of the underlying logic that return values to the ConcertController, which then returns them to the EventController, which then gives them to the SceneController to then display them on screen for the user to see.